Refactor clint rules: split rules.py into individual rule files
- Split large rules.py (559 lines) into modular structure with one rule per file
- Created rules/ directory with separate files for each rule class
- Moved base Rule class to rules/base.py
- Created utils.py for shared utility functions
- Updated __init__.py to export all rules for backward compatibility
- Removed unnecessary class-level docstrings from rule files
- Improved code organization and maintainability
