+++ dev/clint/src/clint/rules/docstring_param_order.py
@@ -1,0 +1,9 @@
+from clint.rules.base import Rule
+
+
+class DocstringParamOrder(Rule):
+    def __init__(self, params: list[str]) -> None:
+        self.params = params
+
+    def _message(self) -> str:
+        return f"Unordered parameters in docstring: {self.params}"
