+++ dev/clint/src/clint/rules/empty_notebook_cell.py
@@ -1,0 +1,6 @@
+from clint.rules.base import Rule
+
+
+class EmptyNotebookCell(Rule):
+    def _message(self) -> str:
+        return "Empty notebook cell. Remove it or add some content."
