+++ dev/clint/src/clint/rules/example_syntax_error.py
@@ -1,0 +1,6 @@
+from clint.rules.base import Rule
+
+
+class ExampleSyntaxError(Rule):
+    def _message(self) -> str:
+        return "This example has a syntax error."
