+++ dev/clint/src/clint/rules/implicit_optional.py
@@ -1,0 +1,48 @@
+import ast
+
+from clint.rules.base import Rule
+
+
+class ImplicitOptional(Rule):
+    def _message(self) -> str:
+        return "Use `Optional` if default value is `None`"
+
+    @staticmethod
+    def check(node: ast.AnnAssign) -> bool:
+        """
+        Returns True if the value to assign is `None` but the type annotation is
+        not `Optional[...]` or `... | None`. For example: `a: int = None`.
+        """
+        return ImplicitOptional._is_none(node.value) and not (
+            ImplicitOptional._is_optional(node.annotation)
+            or ImplicitOptional._is_bitor_none(node.annotation)
+        )
+
+    @staticmethod
+    def _is_optional(ann: ast.AST) -> bool:
+        """
+        Returns True if `ann` looks like `Optional[...]`.
+        """
+        return (
+            isinstance(ann, ast.Subscript)
+            and isinstance(ann.value, ast.Name)
+            and ann.value.id == "Optional"
+        )
+
+    @staticmethod
+    def _is_bitor_none(ann: ast.AST) -> bool:
+        """
+        Returns True if `ann` looks like `... | None`.
+        """
+        return (
+            isinstance(ann, ast.BinOp)
+            and isinstance(ann.op, ast.BitOr)
+            and (isinstance(ann.right, ast.Constant) and ann.right.value is None)
+        )
+
+    @staticmethod
+    def _is_none(value: ast.AnnAssign) -> bool:
+        """
+        Returns True if `value` represents `None`.
+        """
+        return isinstance(value, ast.Constant) and value.value is None
