+++ dev/clint/src/clint/rules/lazy_builtin_import.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from clint.rules.base import Rule
+
+
+class LazyBuiltinImport(Rule):
+    def _message(self) -> str:
+        return "Builtin modules must be imported at the top level."
