+++ dev/clint/src/clint/rules/markdown_link.py
@@ -1,0 +1,9 @@
+from clint.rules.base import Rule
+
+
+class MarkdownLink(Rule):
+    def _message(self) -> str:
+        return (
+            "Markdown link is not supported in docstring. "
+            "Use reST link instead (e.g., `Link text <link URL>`_)."
+        )
