+++ dev/clint/src/clint/rules/mlflow_class_name.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from clint.rules.base import Rule
+
+
+class MlflowClassName(Rule):
+    def _message(self) -> str:
+        return "Should use `Mlflow` in class name, not `MLflow` or `MLFlow`."
