+++ dev/clint/src/clint/rules/no_rst.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from clint.rules.base import Rule
+
+
+class NoRst(Rule):
+    def _message(self) -> str:
+        return "Do not use RST style. Use Google style instead."
