+++ dev/clint/src/clint/rules/test_name_typo.py
@@ -1,0 +1,6 @@
+from clint.rules.base import Rule
+
+
+class TestNameTypo(Rule):
+    def _message(self) -> str:
+        return "This function looks like a test, but its name does not start with 'test_'."
