+++ dev/clint/src/clint/rules/unnamed_thread.py
@@ -1,0 +1,23 @@
+import ast
+
+from clint.resolver import Resolver
+from clint.rules.base import Rule
+
+
+class UnnamedThread(Rule):
+    def _message(self) -> str:
+        return (
+            "`threading.Thread()` must be called with a `name` argument to improve debugging "
+            "and traceability of thread-related issues."
+        )
+
+    @staticmethod
+    def check(node: ast.Call, resolver: Resolver) -> bool:
+        """
+        Returns True if the call is threading.Thread() without a name parameter.
+        """
+        return (
+            (resolved := resolver.resolve(node))
+            and resolved == ["threading", "Thread"]
+            and not any(keyword.arg == "name" for keyword in node.keywords)
+        )
